Add : 게시글 작성자만 수정, 삭제 가능하도록 구현 #2

+++ src/controller/postController.js
@@ -18,23 +18,34 @@ export async function getPost(req, res) {
 }
 
 export async function createPost(req, res) {
-  const { content, username, name } = req.body;
-  const newPost = await postRepo.create(content, username, name);
+  const { content } = req.body;
+  const newPost = await postRepo.create(content, req.userId);
   res.status(201).json(newPost);
 }
 
 export async function updatePost(req, res) {
   const { id } = req.params;
   const { content } = req.body;
-  const post = await postRepo.update(id, content);
+  const post = await postRepo.getById(id);
   if (!post) {
     return res.status(404).json({ msg: "존재하지 않는 게시글입니다." });
   }
-  res.status(200).json(post);
+  if (post.userId !== req.userId) {
+    return res.status(403).json({ msg: "권한이 없습니다." });
+  }
+  const updatedPost = await postRepo.update(id, content);
+  res.status(200).json(updatedPost);
 }
 
 export async function deletePost(req, res) {
   const { id } = req.params;
+  const post = await postRepo.getById(id);
+  if (!post) {
+    return res.status(404).json({ msg: "존재하지 않는 게시글입니다." });
+  }
+  if (post.userId !== req.userId) {
+    return res.status(403).json({ msg: "권한이 없습니다." });
+  }
   await postRepo.remove(id);
   res.sendStatus(204);
 }
